Split `tonic::api::points_common` module (#5865)
