+++ src/tonic/api/mod.rs
@@ -1,11 +1,13 @@
 pub mod collections_api;
-mod collections_common;
 pub mod collections_internal_api;
 pub mod points_api;
-mod points_common;
 pub mod points_internal_api;
 pub mod raft_api;
 pub mod snapshots_api;
+
+mod collections_common;
+mod query_common;
+mod update_common;
 
 use collection::operations::validation;
 use tonic::Status;
